requirements.txt
fastapi
uvicorn
tensorflow
numpy
pandas
scikit-learn
joblib
matplotlib
python-binance